ssh2: Add timeout parameter for ssh2 server
The ssh2 server disconnects the client after some time. This time is now configurable. However with this commit it does not work yet. The problem to be solved is to get to the created session (IoSession, ServerSession) and change the bothTimeouts value.
This patch also addresses the history file path that was wrong in the admin properties.

Ticket:
Acked-by: none yet
Target: trunk
Require-book: yes
Require-notes: yes